Organize calculation models
