+++ src/models/beam_model.py
